clear_user_session_maps -> clear_expired_sessions
Now clears both Django Session and UserSessionMap objects.
